feat: Let the errors from package loading be inspected
Return an array of errors as a second argument of the callback.
Promise an object with packages and errors.

BREAKING CHANGE: The promised content is not the same as the first callback argument. It is an object with properties `packages` and `errors`. These properties correspond with the first and second arguments of the callback.